+++ langs/csharp/src/VMx/Aggregates/AggregateVM5.cs
@@ -223,14 +223,14 @@ public sealed class AggregateVM5<VM1, VM2, VM3, VM4, VM5> : ComponentVMBase, IAg
         /// </summary>
         public AggregateVM5<VM1, VM2, VM3, VM4, VM5> Build()
         {
-            if (_name is null) throw new BuilderValidationException("Name");
-            if (_hub is null) throw new BuilderValidationException("Hub");
-            if (_dispatcher is null) throw new BuilderValidationException("Dispatcher");
-            if (_factory1 is null) throw new BuilderValidationException("Component1");
-            if (_factory2 is null) throw new BuilderValidationException("Component2");
-            if (_factory3 is null) throw new BuilderValidationException("Component3");
-            if (_factory4 is null) throw new BuilderValidationException("Component4");
-            if (_factory5 is null) throw new BuilderValidationException("Component5");
+            BuilderValidationException.Require(_name, "Name");
+            BuilderValidationException.Require(_hub, "Hub");
+            BuilderValidationException.Require(_dispatcher, "Dispatcher");
+            BuilderValidationException.Require(_factory1, "Component1");
+            BuilderValidationException.Require(_factory2, "Component2");
+            BuilderValidationException.Require(_factory3, "Component3");
+            BuilderValidationException.Require(_factory4, "Component4");
+            BuilderValidationException.Require(_factory5, "Component5");
 
             return new AggregateVM5<VM1, VM2, VM3, VM4, VM5>(
                 _name, _hint, _hub, _dispatcher,